Create Football Prediction
This app I have created a function predict_outcome that takes in the user's football data and returns predictions from multiple regression models. The main function main checks if the user has made a payment of N$20 for the current week using the check_payment function from a payment system library. If the payment has been made, the app calls the predict_outcome function and prints the predictions. If the payment has not been made, the app prompts the user to make a payment.

There are different payment system library used by this app is  such as paypal, square and other payment system that I choose to integrate with your app. Additionally, I may need to modify this code to fit your specific requirements and architecture in the future as it is required.